requirements.txt
fastapi==0.115.11
passlib==1.7.4
pydantic==2.10.6
python-dotenv==1.0.1
python_jose==3.4.0
SQLAlchemy==2.0.39
